# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface AuthUser {
  userId: string;
  email: string;
  firstName: string;
  lastName: string;
  address?: string;
  city?: string;
  state?: string;
  stateId?: string;
  zipCode?: string;
  phoneNumber?: string;
  profileImage?: string;
  status?: string;
  /** Raw subscription expiry date from the legacy backend; empty/zero when there is no active subscription. */
  subscribeExpiryDate?: string;
  /** Whether the account currently has an active paid subscription. */
  isSubscribed?: boolean;
}

export interface AuthResponse {
  /** @minLength 1 */
  authToken: string;
  user: AuthUser;
}

export interface DeleteAccountResult {
  success: boolean;
}

export interface StaticPage {
  title: string;
  contentHtml: string;
}

export interface AuthLoginInput {
  /** @minLength 1 */
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface AuthSignUpInput {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  /** @minLength 1 */
  email: string;
  /** @minLength 1 */
  password: string;
  /** @minLength 1 */
  stateId: string;
  /** @minLength 1 */
  city: string;
  /** @minLength 1 */
  zipCode: string;
  phone?: string;
}

/**
 * The short-lived, single-use code handed to the app on its return URL at the end of the server-mediated Google OAuth flow. The server verifies the identity with Google directly, so the app never sends provider data.
 */
export interface AuthGoogleExchangeInput {
  /** @minLength 1 */
  code: string;
}

/**
 * Completes Google sign-up for a verified-but-unlinked identity. `code` is the short-lived registration ticket handed to the app on its return URL; `city` and `zipCode` are the minimum the legacy backend requires to create the account. The verified email and name are held server-side against the ticket, so they are never sent by the app.
 */
export interface AuthGoogleRegisterInput {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  city: string;
  /** @minLength 1 */
  zipCode: string;
}

/**
 * Partial profile update for the authenticated user. `state` is the 2-letter US abbreviation; the server maps it to the legacy numeric state id. All fields are optional so partial updates are allowed.
 */
export interface AuthProfileUpdateInput {
  firstName?: string;
  lastName?: string;
  address?: string;
  city?: string;
  state?: string;
  zipCode?: string;
  phone?: string;
}

export interface Election {
  id: string;
  name: string;
  electionDay: string;
  /** @nullable */
  ocdDivisionId?: string | null;
}

export interface Elections {
  elections: Election[];
}

export interface SocialChannel {
  type: string;
  id: string;
}

export interface Candidate {
  name: string;
  /** @nullable */
  party?: string | null;
  /** @nullable */
  candidateUrl?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  channels: SocialChannel[];
}

export interface Contest {
  type: string;
  /** @nullable */
  office?: string | null;
  /** @nullable */
  level?: string | null;
  /** @nullable */
  districtScope?: string | null;
  candidates: Candidate[];
}

export interface CorrespondenceAddress {
  /** @nullable */
  line1?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  zip?: string | null;
}

export interface StateAdministration {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  electionInfoUrl?: string | null;
  /** @nullable */
  votingLocationFinderUrl?: string | null;
  /** @nullable */
  ballotInfoUrl?: string | null;
  correspondenceAddress?: CorrespondenceAddress;
}

export interface PollingLocation {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  hours?: string | null;
}

export interface VoterInfo {
  address: string;
  /** @nullable */
  electionId?: string | null;
  state?: StateAdministration;
  contests: Contest[];
  pollingLocations: PollingLocation[];
}

export interface PollOption {
  id: string;
  label: string;
  votes: number;
}

export interface Poll {
  id: string;
  question: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  category?: string | null;
  totalVotes: number;
  options: PollOption[];
}

export interface PollList {
  polls: Poll[];
}

export interface PollVoteInput {
  /** @minLength 1 */
  optionId: string;
}

export interface PollResultOption {
  id: string;
  label: string;
  votes: number;
  percentage: number;
}

export interface PollResults {
  pollId: string;
  question: string;
  totalVotes: number;
  options: PollResultOption[];
}

export interface HomeSection {
  id: string;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  route?: string | null;
}

export interface HomeData {
  appName: string;
  environment: string;
  newsUrl: string;
  lastUpdated: string;
  sections: HomeSection[];
}

export interface NewsArticle {
  id: string;
  title: string;
  description: string;
  source: string;
  url: string;
  publishedAt: string;
  /** Plain-text article body for in-app reading (paragraphs separated by blank lines). */
  content?: string;
  /** Lead image URL extracted from the feed item, if any. */
  imageUrl?: string;
  /** Article author/byline, if provided by the feed. */
  author?: string;
}

export interface NewsResponse {
  newsUrl: string;
  title: string;
  description: string;
  source: string;
  articles: NewsArticle[];
}

export interface Quiz {
  id: string;
  title: string;
  /** @nullable */
  description?: string | null;
  category: string;
  questionCount: number;
}

export interface QuizList {
  quizzes: Quiz[];
}

export interface CivicsQuizQuestion {
  id: string;
  question: string;
  options: string[];
  answerIndex: number;
  explanation: string;
}

export interface CivicsQuiz {
  id: string;
  title: string;
  description: string;
  /** @minItems 1 */
  questions: CivicsQuizQuestion[];
}

export interface QuizProgressInput {
  /** @minLength 1 */
  quizId: string;
  userId?: string;
  score?: number;
  completed: boolean;
}

export interface QuizProgressResult {
  success: boolean;
  quizId: string;
  recordedAt: string;
}

export interface QuestionnaireQuestion {
  id: string;
  prompt: string;
  options: string[];
}

export interface Questionnaire {
  id: string;
  title: string;
  /** @nullable */
  description?: string | null;
  questions: QuestionnaireQuestion[];
}

export interface QuestionnaireList {
  questionnaires: Questionnaire[];
}

export interface TerminologyItem {
  id: string;
  term: string;
  definition: string;
  /** @nullable */
  category?: string | null;
}

export interface TerminologyList {
  terms: TerminologyItem[];
}

export interface Party {
  id: string;
  name: string;
  /** @nullable */
  abbreviation?: string | null;
  description: string;
  /** @nullable */
  foundedYear?: number | null;
  /** @nullable */
  website?: string | null;
}

export interface PartyList {
  parties: Party[];
}

export type GetVoterInfoParams = {
address: string;
electionId?: string;
};