# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Authenticate a user against the legacy server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLoginInput"
      responses:
        "200":
          description: Authenticated user with token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Register a new user against the legacy server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthSignUpInput"
      responses:
        "200":
          description: Newly registered user with token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid request or registration rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /auth/profile:
    post:
      operationId: updateProfile
      tags: [auth]
      summary: Update the authenticated user's profile on the legacy server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthProfileUpdateInput"
      responses:
        "200":
          description: Updated user with token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid request or update rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Missing or rejected auth token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/account:
    delete:
      operationId: deleteAccount
      tags: [auth]
      summary: Permanently delete the authenticated user's account
      description: >-
        Deletes the signed-in user's account on the legacy backend. The user is
        identified by the forwarded AUTHTOKEN, so no body is required. The app
        clears the local session after a successful deletion.
      responses:
        "200":
          description: Account deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteAccountResult"
        "401":
          description: Missing or rejected auth token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    AuthUser:
      type: object
      properties:
        userId:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        stateId:
          type: string
        zipCode:
          type: string
        phoneNumber:
          type: string
        profileImage:
          type: string
        status:
          type: string
        subscribeExpiryDate:
          type: string
          description: Raw subscription expiry date from the legacy backend; empty/zero when there is no active subscription.
        isSubscribed:
          type: boolean
          description: Whether the account currently has an active paid subscription.
      required:
        - userId
        - email
        - firstName
        - lastName
    AuthResponse:
      type: object
      properties:
        authToken:
          type: string
          minLength: 1
        user:
          $ref: "#/components/schemas/AuthUser"
      required:
        - authToken
        - user
    DeleteAccountResult:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    AuthLoginInput:
      type: object
      properties:
        email:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
      required:
        - email
        - password
    AuthSignUpInput:
      type: object
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        email:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
        stateId:
          type: string
          minLength: 1
        city:
          type: string
          minLength: 1
        zipCode:
          type: string
          minLength: 1
        phone:
          type: string
      required:
        - firstName
        - lastName
        - email
        - password
        - stateId
        - city
        - zipCode
...
    AuthProfileUpdateInput:
      type: object
      description: >-
        Partial profile update for the authenticated user. `state` is the
        2-letter US abbreviation; the server maps it to the legacy numeric
        state id. All fields are optional so partial updates are allowed.
      properties:
        firstName:
          type: string
        lastName:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        phone:
          type: string
...