# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthUser {
  userId: string;
  email: string;
  firstName: string;
  lastName: string;
  address?: string;
  city?: string;
  state?: string;
  stateId?: string;
  zipCode?: string;
  phoneNumber?: string;
  profileImage?: string;
  status?: string;
  /** Raw subscription expiry date from the legacy backend; empty/zero when there is no active subscription. */
  subscribeExpiryDate?: string;
  /** Whether the account currently has an active paid subscription. */
  isSubscribed?: boolean;
}